requirements.txt
pandas==1.3.3
Pillow==9.0.1
scikit_learn==1.0.2
torch==1.9.1+cpu
torchvision==0.10.1+cpu
wandb==0.12.7
